Refactor resends logic into src/Resender.
